Added scene manager
