+++ src/engine/core/Engine.ts
@@ -1,5 +1,6 @@
 import Phaser from 'phaser';
 import Scene from './Scene';
+import SceneManager from './SceneManager';
 
 export default class Engine extends Phaser.Scene {
   private _width: number;
@@ -10,6 +11,8 @@ export default class Engine extends Phaser.Scene {
   private _ready_resolver: (() => void) | null = null;
   private _ready_promise: Promise<void>;
 
+  private _scene_manager: SceneManager = new SceneManager(this);
+
   // [OK] Make sure the phaser scene is ready before doing things in the Scene
   // Assets
   // Systems
@@ -17,8 +20,8 @@ export default class Engine extends Phaser.Scene {
   // [OK] Scene transitions/fade in/fade out (handled manually for now)
   // - Stop scene
   // - Pause scene
-  // Add scene manager
-  // ** Pass scene instead of world to systems
+  // [OK] Add scene manager
+  // [OK] Pass scene instead of world to systems
   // Serialization
   // Testing
   // Comments
@@ -45,73 +48,8 @@ export default class Engine extends Phaser.Scene {
     return this._ready_promise;
   }
 
-  registerScene(key: string, scene: Scene): void {
-    if (!this._registered_scenes.has(key)) {
-      scene.initialize(this);
-      scene.onRegister();
-      this._registered_scenes.set(key, scene);
-    }
-  }
-
-  startScene(key: string): void {
-    const scene = this._registered_scenes.get(key);
-    if (!scene) {
-      throw new Error(`Cannot start scene ${key}, it's not registered.`);
-    }
-
-    if (this._active_scenes.has(key)) {
-      console.warn(`Scene ${key} is already active.`);
-      return;
-    }
-
-    console.log(`Starting scene: ${key}`);
-    scene.onStart();
-    this._active_scenes.set(key, scene);
-  }
-
-  stopScene(key: string): void {
-    const scene = this._active_scenes.get(key);
-    if (!scene) {
-      throw new Error(`Scene ${key} is not active.`);
-    }
-
-    console.log(`Stopping scene: ${key}`);
-    scene.onExit();
-    const drawSystem = scene.world.getDrawSystem();
-    if (drawSystem) {
-      drawSystem.cleanupAll();
-    }
-    this._active_scenes.delete(key);
-  }
-
-  pauseScene(key: string): void {
-    const scene = this._active_scenes.get(key);
-    if (!scene) {
-      throw new Error(`Scene ${key} is not active and cannot be paused.`);
-    }
-
-    console.log(`Pausing scene: ${key}`);
-    scene.onPause();
-    const drawSystem = scene.world.getDrawSystem();
-    if (drawSystem) {
-      drawSystem.cleanupAll();
-    }
-    this._active_scenes.delete(key);
-  }
-
-  resumeScene(key: string): void {
-    const scene = this._registered_scenes.get(key);
-    if (!scene) {
-      throw new Error(`Cannot resume scene ${key}, it's not registered.`);
-    }
-
-    if (this._active_scenes.has(key)) {
-      throw new Error(`Scene ${key} is already active.`);
-    }
-
-    console.log(`Resuming scene: ${key}`);
-    scene.onResume();
-    this._active_scenes.set(key, scene);
+  getSceneManager(): SceneManager {
+    return this._scene_manager;
   }
 
   preload(): void {
@@ -127,8 +65,6 @@ export default class Engine extends Phaser.Scene {
   }
 
   update(time: number, delta: number): void {
-    this._active_scenes.forEach((scene) => {
-      scene.onUpdate(time, delta);
-    });
+    this._scene_manager.updateActiveScenes(time, delta);
   }
 }
